requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
discord.py==1.0.0a1477+g184c430
idna==2.7
idna-ssl==1.1.0
matrix-client==0.3.2
multidict==4.4.0
pkg-resources==0.0.0
requests==2.19.1
urllib3==1.23
websockets==5.0.1
yarl==1.2.6
